+++ lib/chromatin-data.js
@@ -1,5 +1,99 @@
 import _ from "lodash";
-import { ChromatinStateColor } from "../components/chromatin-state-facets";
+
+export const ChromatinStateColor = {
+  "Active TSS": {
+    tailwind: "bg-rose-600",
+    hex: "#fbbf24",
+  },
+  "Flanking TSS": {
+    tailwind: "bg-orange-500",
+    hex: "#f97316",
+  },
+  "Flanking TSS downstream": {
+    tailwind: "bg-orange-500",
+    hex: "#f97316",
+  },
+  "Flanking TSS upstream": {
+    tailwind: "bg-orange-500",
+    hex: "#f97316",
+  },
+  "Active enhancer 1": {
+    tailwind: "bg-amber-400",
+    hex: "#fbbf24",
+  },
+  "Active enhancer 2": {
+    tailwind: "bg-amber-400",
+    hex: "#fbbf24",
+  },
+  "Weak enhancer": {
+    tailwind: "bg-yellow-300",
+    hex: "#fde047",
+  },
+  "Genic enhancer 1": {
+    tailwind: "bg-lime-400",
+    hex: "#a3e635",
+  },
+  "Genic enhancer 2": {
+    tailwind: "bg-lime-400",
+    hex: "#a3e635",
+  },
+  "Strong transcription": {
+    tailwind: "bg-green-600",
+    hex: "#16a34a",
+  },
+  "Weak transcription": {
+    tailwind: "bg-green-600",
+    hex: "#16a34a",
+  },
+  "Bivalent/Poised TSS": {
+    tailwind: "bg-rose-400",
+    hex: "#fb7185",
+  },
+  "Bivalent enhancer": {
+    tailwind: "bg-yellow-500",
+    hex: "#eab308",
+  },
+  "ZNF genes & repeats": {
+    tailwind: "bg-emerald-400",
+    hex: "#34d399",
+  },
+  "Weak Repressed PolyComb": {
+    tailwind: "bg-gray-400",
+    hex: "#a8a29e",
+  },
+  "Repressed PolyComb": {
+    tailwind: "bg-gray-400",
+    hex: "#a8a29e",
+  },
+  Heterochromatin: {
+    tailwind: "bg-indigo-400",
+    hex: "#818cf8",
+  },
+  "Quiescent/Low": {
+    tailwind: "bg-gray-300",
+    hex: "#d6d3d1",
+  },
+  Enhancers: {
+    tailwind: "bg-yellow-300",
+    hex: "#fde047",
+  },
+  "Flanking Bivalent TSS/Enh": {
+    tailwind: "bg-red-300",
+    hex: "#fca5a5",
+  },
+  "Flanking Active TSS": {
+    tailwind: "bg-orange-500",
+    hex: "#f97316",
+  },
+  "Transcr. at gene 5' and 3'": {
+    tailwind: "bg-green-300",
+    hex: "#86efac",
+  },
+  "Genic enhancers": {
+    tailwind: "bg-lime-400",
+    hex: "#a3e635",
+  },
+};
 
 export const ASSOCIATED_ORGAN_MAP = {
   breast: ["mammary gland"],
@@ -178,6 +272,11 @@ export function isLetter(c) {
   return c.toLowerCase() !== c.toUpperCase();
 }
 
+/**
+ * It filters and cleans up the original data and returns a cleaned data array.
+ * @param {Array} data The original data got from GDS
+ * @returns The cleaned data array
+ */
 export function getChromatinData(data) {
   const filteredData = data.filter((d) => d.method === "chromatin state");
   const chromatinData = [];
@@ -204,6 +303,15 @@ export function getChromatinData(data) {
   return chromatinData;
 }
 
+/**
+ * It generates the facets data used for BodyMap.
+ * The data is a object contains keys of different organ slims.
+ * for each organ slim key, the value is an array of the states the organ slim has,
+ * sorded by state activity.
+ * @param {array} data to get organ facets
+ * @param {string} assembly for the data
+ * @returns the organ facets data
+ */
 export function getOrganFacets(data, assembly) {
   const order =
     assembly === "hg19"
@@ -239,7 +347,14 @@ export function getOrganFacets(data, assembly) {
   return sorted;
 }
 
-export function getDisplayedStatesFacets(data, assembly) {
+/**
+ * It generates the facets data used for ChromatinStateFacets.
+ * The data is an array of chromatin states sorted by states activity.
+ * @param {array} data to get state facets
+ * @param {*} assembly for the data
+ * @returns the state facets data
+ */
+export function getStatesFacets(data, assembly) {
   const order =
     assembly === "hg19"
       ? SORTED_CHROMATIN_STATES_HG19
@@ -259,34 +374,16 @@ export function getDisplayedStatesFacets(data, assembly) {
   return sortedStates;
 }
 
-export function getEnabledBodyMapFilters(data) {
-  const bodyMapFilters = data.reduce((acc, dataset) => {
-    const organsForDataset = dataset.organ.split(", ").filter((e) => {
-      return e;
-    });
-    organsForDataset.forEach((organ) => {
-      if (!acc.includes(organ)) {
-        acc.push(organ);
-      }
-    });
-    return acc;
-  }, []);
-  return bodyMapFilters;
-}
-
-export function getFillColorHex(facets, organ) {
-  const fill =
-    organ in facets ? ChromatinStateColor[facets[organ][0]].hex : "white";
-
-  return fill;
-}
-
-export function getFillColorTailwind(facets, organ) {
-  const fill =
-    organ in facets ? ChromatinStateColor[facets[organ][0]].tailwind : "white";
-  return fill;
-}
-
+/**
+ * It generates the facets data used for ChromatinBiosampleFacets.
+ * The data is an object contains keys of different biosample.
+ * Each biosample has those info: biosample name, organ slims, and states
+ * associated to the biosample sorted by states activity. For the states,
+ * it contains the state names as keys and the number of dataset for each state as values.
+ * @param {array} data to get state facets
+ * @param {*} assembly for the data
+ * @returns the state facets data
+ */
 export function getBiosampleFacets(data, assembly) {
   const order =
     assembly === "hg19"
@@ -321,6 +418,15 @@ export function getBiosampleFacets(data, assembly) {
   return filteredFacets;
 }
 
+/**
+ * It returns the filtered data, filtered by state filters,
+ * organ filters and biosample filters.
+ * @param {array} data the unfiltered data
+ * @param {*} stateFilters
+ * @param {*} organFilters
+ * @param {*} biosampleFilters
+ * @returns the filtered data.
+ */
 export function getFilteredChromatinData(
   data,
   stateFilters,
@@ -354,4 +460,23 @@ export function getFilteredChromatinData(
     });
   }
   return filteredData;
+}
+
+/**
+ * Return color in hex
+ */
+export function getFillColorHex(facets, organ) {
+  const fill =
+    organ in facets ? ChromatinStateColor[facets[organ][0]].hex : "white";
+
+  return fill;
+}
+
+/**
+ * Return color in tailwind style
+ */
+export function getFillColorTailwind(facets, organ) {
+  const fill =
+    organ in facets ? ChromatinStateColor[facets[organ][0]].tailwind : null;
+  return fill;
 }
