+++ lib/general.js
@@ -112,7 +112,3 @@ export function sanitizedString(inputString) {
     .replace(/[*?()+[\]\\/]/g, "") // remove certain special characters (these cause console errors)
     .replace("&", "");
 }
-
-export function classString(inputString) {
-  return inputString.replace(/[^A-Za-z]+/g, "");
-}
